product feature

+++ src/main/java/com/vichika/ecommercesystem/product/ProductController.java
@@ -19,14 +19,14 @@ public class ProductController {
     private final ProductService productService;
 
     @GetMapping
-    public ResponseEntity<APIResponse<PageResponse<ProductResponse>>> getAll(@RequestParam(required = false) Byte id,
-                                                                            @RequestParam(required = false) String name,
-                                                                            @RequestParam(required = false) String code,
-                                                                            @RequestParam(required = false) Integer categoryId,
-                                                                            @RequestParam(required = false) String sortBy,
-                                                                            @RequestParam(required = false) String sortAs,
-                                                                            @RequestParam(required = false,defaultValue = "1") Integer page,
-                                                                            @RequestParam(required = false,defaultValue = "5") Integer size){
+    public ResponseEntity<APIResponse<PageResponse<ProductResponse>>> getAll(@RequestParam(required = false) Long id,
+                                                                             @RequestParam(required = false) String name,
+                                                                             @RequestParam(required = false) String code,
+                                                                             @RequestParam(required = false) Byte categoryId,
+                                                                             @RequestParam(required = false) String sortBy,
+                                                                             @RequestParam(required = false) String sortAs,
+                                                                             @RequestParam(required = false,defaultValue = "1") Integer page,
+                                                                             @RequestParam(required = false,defaultValue = "5") Integer size){
         return ResponseEntity.ok(APIResponse.ok(productService.getAllProduct(id,name,code,categoryId,sortBy,sortAs,page,size)));
     }
