+++ src/main/java/com/vichika/ecommercesystem/product/service/ProductService.java
@@ -6,7 +6,7 @@ import com.vichika.ecommercesystem.product.dto.response.ProductResponse;
 import org.springframework.web.multipart.MultipartFile;
 
 public interface ProductService {
-    PageResponse<ProductResponse> getAllProduct(Byte id, String name, String code, Integer categoryId, String sortBy, String sortAs, Integer page, Integer size);
+    PageResponse<ProductResponse> getAllProduct(Long id, String name, String code, Byte categoryId, String sortBy, String sortAs, Integer page, Integer size);
 
     ProductResponse createProduct(ProductRequest request);
